+++ Project.toml
@@ -16,8 +16,15 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Suppressor = "fd094767-a336-5f1f-9728-57cf17d0bbfb"
 YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 
+[weakdeps]
+PlutoStaticHTML = "359b1769-a58e-495b-9770-312e911026ad"
+
+[extensions]
+PlutoNotebook = "PlutoStaticHTML"
+
 [compat]
 Documenter = "0.22, 0.23, 0.24, 0.25, 0.26, 0.27"
+PlutoStaticHTML = "6"
 Requires = "1.3"
 FileIO = "1"
 HTTP = "0.6, 0.7, 0.8, 0.9, 1"
@@ -30,6 +37,7 @@ YAML = "0.3, 0.4"
 julia = "1.6"
 
 [extras]
+PlutoStaticHTML = "359b1769-a58e-495b-9770-312e911026ad"
 ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
 ImageShow = "4e3cecfd-b093-5904-9786-8bbb286a6a31"
 ReferenceTests = "324d217c-45ce-50fc-942e-d289b448e8cf"
@@ -37,4 +45,4 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TestImages = "5e47fb64-e119-507b-a336-dd2b206d9990"
 
 [targets]
-test = ["ImageMagick", "ImageShow", "ReferenceTests", "Test", "TestImages"]
+test = ["ImageMagick", "ImageShow", "ReferenceTests", "Test", "TestImages", "PlutoStaticHTML"]
