+++ src/DemoCards.jl
@@ -6,7 +6,6 @@ using Mustache
 using Literate
 using ImageCore
 using FileIO, JSON, YAML
-using Requires
 using Suppressor # suppress log generated by 3rd party tools, e.g., Literate
 import HTTP
 using Documenter
@@ -48,10 +47,14 @@ include("preview.jl")
 
 export makedemos, cardtheme, preview_demos
 
+if !isdefined(Base, :get_extension)
+using Requires
+end
+
 function __init__()
-    @require Pluto="c3e4b0f8-55cb-11ea-2926-15256bba5781" begin
-      @require PlutoStaticHTML = "359b1769-a58e-495b-9770-312e911026ad" include("types/pluto.jl")
-    end
+  @static if !isdefined(Base, :get_extension)
+    @require PlutoStaticHTML = "359b1769-a58e-495b-9770-312e911026ad" include("../ext/PlutoNotebook.jl")
+  end
 end
 
 """
