+++ src/types/card.jl
@@ -126,3 +126,4 @@ end
 
 include("markdown.jl")
 include("julia.jl")
+include("pluto.jl")
